+++ index.ts
@@ -841,3 +841,4 @@ export {
 } from './src/util/flight/flightVisualizer.js';
 export { getTileLoadFunction } from './src/layer/openlayers/loadFunctionHelpers.js';
 export { default as ModelFill } from './src/style/modelFill.js';
+export { renderTemplate } from './src/util/vcsTemplate.js';